% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_st_end_centr_arms.R
\name{chr_st_en_etc}
\alias{chr_st_en_etc}
\title{internal cnvgeaRs function}
\usage{
chr_st_en_etc()
}
\description{
Generates chromosome start, end centrosome location, as well as chromosomal
arm start end datatsets in three different assemblies.
}
