% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_uniform.R
\name{chr_uniform}
\alias{chr_uniform}
\title{Uniform chromosome notation}
\usage{
chr_uniform(DT_in)
}
\arguments{
\item{DT_in, }{a \code{data.table} with a columns named "chr"}
}
\value{
the same \code{data.table} in input with the "chr" uniformed to the
  notation "1", "2", ... , "23", "24" for the chromosomes 1:22, X and Y
}
\description{
This is a function for internal use in the package, it handles
the standardize process of chromosome notation within the other functions.
}
\examples{
DT <- data.table::data.table(chr= c("chr1", "chrX", "chr20"))
DT <- chr_uniform(DT)
DT
}
