% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_calls.R
\name{cleaning_filter}
\alias{cleaning_filter}
\title{Filter CNVs calls based on several parameters}
\usage{
cleaning_filter(
  results,
  min_len = 10000,
  min_NP = 10,
  blacklists = NULL,
  blacklist_samples = NULL,
  blacklist_chrs = NULL
)
}
\arguments{
\item{results, }{a \code{data.table}, the output of
\code{\link{read_results}}.}

\item{min_len, }{minimum CNVs length, any shorter event will be filtered out.
Default is 10000.}

\item{min_NP, }{minimum CNVs points, any shorter event will be filtered out.
Default is 10.}

\item{blacklists, }{blacklist, a \code{list} of \code{data.table} or \code{data.frame}
containing at least the following columns: "chr", "start", "end". Any event
in a blacklists' region will be filtered out.}

\item{blacklist_samples, }{character vector containing samples ID to filter
out.}

\item{blacklist_chrs, }{character vector containing chromosomes names in the
package format, 1:22 for autosomes and 23 24 for chr X and Y.}
}
\value{
the \code{CNVresults} object \code{results} after the selected filters
  have been applied.
}
\description{
\code{cleaning_filter} "clean" a CNV calls dataset based on measures such as
length, number of calls per sample and more.
}
\details{
This function can be used together with \code{\link{summary}} in order
to clean the dataset from possible noise and unwanted calls. It is generally
recommended to briefly explore the data using \code{\link{summary}} and
then proceeding to filter out any unwanted group of events. Mandatory
arguments of the function are "results" and "min_len"/"min_NP", default
values are the authors suggested minimal filtering step, however its quite
common to filter anything shorter than 10 or even 50 kb, and/or any call made
by less than 10 points. The use of blacklist of any kind is optional and
should be done with caution, as it can filter potential biologically relevant
events. Over-segmented samples the user wishes to exclude can be specified via
\code{blacklist_samples}. Immunoglobulin regions can be generated with the function
\code{\link{immuno_regions}}, while telomeric and/or centromeric regions can be
obtained with \code{\link{telom_centrom}}.
}
\examples{

DT <- cleaning_filter(penn_22, min_len = 50000)
}
