% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmops_to_CNVresults.R
\name{cnmops_to_CNVresults}
\alias{cnmops_to_CNVresults}
\title{Convert cn.mops results into \code{CNVgears} format}
\usage{
cnmops_to_CNVresults(cnvs, samples_list, markers)
}
\arguments{
\item{markers, }{a \code{data.table} containing the marker list, the output
\code{\link{read_finalreport_snps}} with \code{DT_type} set to "markers"
or \code{\link{read_NGS_intervals}}.}

\item{sample_list, }{minimal cohort metadata, a \code{data.table} produced by the
function \code{\link{read_metadt}}.}
}
\description{
Convert cn.mops results into \code{CNVgears} format
}
