% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_from_to.R
\name{CNVresults_to_GRanges}
\alias{CNVresults_to_GRanges}
\title{CNVresults to GRanges}
\usage{
CNVresults_to_GRanges(DT)
}
\arguments{
\item{DT}{a \code{CNVresults} object.}
}
\description{
\code{CNVresults_to_GRanges} convert \code{CNVresults} objects into \code{GRanges}
}
\details{
A simple wrapper for the function \code{GenomicRanges::makeGRangesFromDataFrame}.
Retained metadata columns are: sample_ID, GT and meth_ID.
}
