% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVs_inheritance.R
\name{cnvs_inheritance}
\alias{cnvs_inheritance}
\title{Discover good de novo CNVs candidates}
\usage{
cnvs_inheritance(
  sample_list,
  markers,
  results,
  raw_path,
  mmethod = 1,
  alfa = 0.05,
  min_NP = 10,
  pre_fine_screen = TRUE,
  adjust_pval = T,
  reciprocal_overlap = 0.3
)
}
\arguments{
\item{sample_list, }{a \code{data.table}, the output of
\code{\link{read_metadt}}.}

\item{markers, }{a \code{data.table}, the output of
\code{\link{read_NGS_intervals}} or \code{\link{read_finalreport_snps}},
depending on the initial data type.}

\item{results, }{a \code{data.table}, the output of
\code{\link{read_results}}.}

\item{raw_path, }{\code{character}, the path where raw data processed by
\code{\link{read_NGS_raw}} or \code{\link{read_finalreport_raw}}.}
}
\description{
\code{denovo_cnvs} screen CNVs calling results to find de novo candidates,
given a family based study design.
}
\details{
This function
}
