% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uniformation.R
\name{dt_uniform}
\alias{dt_uniform}
\title{Uniform data i/o}
\usage{
dt_uniform(
  DT_in,
  DT_type,
  markers,
  chr_col,
  start_col,
  end_col,
  CN_col,
  LRR_col,
  BAF_col,
  raw_col,
  raw_col_type,
  mark_ID_col,
  sample_ID_col,
  sample_ID,
  sex,
  keep_all = F
)
}
\arguments{
\item{DT_in, }{input data as \code{data.table}.}

\item{DT_type, }{type of input to process, can be either:
"results", "markers_array", "markers_NGS", "raw_array", "raw_NGS". Define
the function's behaviour and the subsequent required parameters.}

\item{markers, }{a \code{data.table} containing the informations on the
markers used in the analysis (SNPs for arrays, intervals for NGS). It must
have a specific format and it is produced by this function via the
\code{DT_type} "markers_array" or "markers_NGS". It is required by the
following \code{DT_type}: "results_array", "results_", "raw_NGS"}

\item{chr_col, }{name of the column containig the chromosome information
in the input data, required by the following \code{DT_type}:
"results", "markers_array", "markers_NGS", "raw_array", "raw_NGS".}

\item{start_col, }{name of the column containig the start information in
the input data, required by the following \code{DT_type}: "results_array",
"results_NGS", "markers_array", "markers_NGS", "raw_array", "raw_NGS".}

\item{end_col, }{name of the column containig the end information in the
input data, required by the following \code{DT_type}: "results_array",
"results_NGS", "markers_NGS", "raw_NGS".}

\item{CN_col, }{name of the column containig the Copy Number information
in the input data, required by the following \code{DT_type}:
"results".}

\item{LRR_col, }{name of the column containig the LRR (Log R Ratio)
information in the input data, required by the following \code{DT_type}:
"raw_array".}

\item{BAF_col, }{name of the column containig the BAF (B Allele
Frquency) information in the input data, required by the following
\code{DT_type}: "raw_array".}

\item{raw_col, }{name of the column containig the Copy Ratio or LRR-like
information in the input data, required by the following \code{DT_type}:
"raw_NGS".}

\item{raw_col_type, }{specify the format of \code{raw_col}, can be
either: "CR", "LRR_like" or "log2CR". Required by the following
\code{DT_type}: "raw_NGS".}

\item{sex, }{Sex of the sample(s) being processed. The format is 0 for
males and 1 for females. Can be a simgle \code{integer} if one single
sample is present or a \code{data.frame} with the columns "ID" and "sex".
It is required by the following \code{DT_type}: "results".}
}
\value{
a \code{data.table}, see the description for the details.
}
\description{
\code{DT_uniform} returns the input (CNVs calling/segmentation results,
markers list or raw data) in a standardized format.
}
\details{
This function handles all the initial data manipulation in order to obtain a
standardized format regardless of the source (array data or NGS), both in
terms of actual data columns presence, and in terms of columns names and
content format and type, it represent the first step of all the possible
analysis. It uses \code{data.table} to speed up the computations on large
dataset, such as the markers on a SNPs array (usually between 500k and
1,500k) or the merged results of a CNV calling/segmentation pipeline on
thousands of samples. Depending on the parameter \code{DT_type} the final
\code{data.table} is consituted of the following columns:
\itemize{
\item \code{chr} (as character), \code{start} (as interger), \code{end} (as
integer), \code{GT} (genotype, as integer, 0 for normal CN, 1 for deletions,
and 2 for duplications), \code{CN} (Copy Number, as integer), \code{segID}
(an univoque integer used as ID of the segment/ CNV), \code{NP} (Number of
Points, SNPs or intervals, that constitute the segment/CNV),
\code{first_P_ID} (integer ID of the fisrt point of the segment/CNV),
\code{last_P_ID}, \code{len} (length of the segment/CNV in bp), \code{sample}
(the sample ID), when \code{DT_type} is "results_array" or "results_NGS";
\item \code{chr} (as character), \code{start} (as interger), \code{end} (as
integer), \code{ID}, \code{name}, when \code{DT_type} is "markers_array" or
"markers_NGS";
\item \code{chr} (as character), \code{start} (as interger),
\code{end} (as integer), \code{ID}, \code{LRR}, \code{BAF}, \code{mark_ID},
when \code{DT_type} is "raw_array";
\item \code{chr} (as character), \code{start} (as interger),
\code{end} (as integer), \code{ID}, raw value depending on
\code{raw_col_type}, when \code{DT_type} is "raw_NGS";
}

This function handles the standardization of most the main data type in this
package: the CNV calling /segmentation results, the markers table, the
finalreports-like tables for array data and the raw LRR-like values for NGS
data. It is intended to be used whithin other functions that handles more
directly the data input/output. Please note that each possible \code{DT_type}
has its own specific requirements both in terms of actual columns in the data
being passed in, and in terms of other objects required in order to being
processed.
}
