% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_load_annotation.R
\name{genic_load}
\alias{genic_load}
\title{Annotate genic load}
\usage{
genic_load(DT_in, biotypes = "protein_coding", mart = NA)
}
\arguments{
\item{DT_in, }{a \code{data.table} consisting of CNVs calling results.}

\item{biotypes, }{character vector of Genecode biotypes, default value is
"protein_coding".}

\item{mart, }{user specified \code{biomaRt::useMart()} object. Used if DT_in
  is assembly (e.g. hg19) require older ensembl releases. If NA the latest
  release available by \code{biomaRt::useMart()} is used.

  This function takes a \code{data.table} containing CNVs calling results
  data as input and add two additional columns representing the genic content
  of each call, i.e. "genes" and "n_genes". The genes considered can be
  changed using the `biotypes` parameter depending on which types of genes is
  the user interested in.}
}
\value{
, a \code{data.table}.
}
\description{
Annotate genic load
}
