% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_annotator.R
\name{genomic_locus}
\alias{genomic_locus}
\title{Rapid genomic locus annotator for CNV calls}
\usage{
genomic_locus(
  DT_in,
  remote_cytobands = TRUE,
  bands,
  assembly = "hg19",
  keep_str_end = TRUE
)
}
\arguments{
\item{DT_in, }{a \code{data.table}, usually \code{\link{read_results}} output.}

\item{remote_cytobands, }{logical, indicates whether the function should
download the CytoBands file or use a local object.}

\item{bands, }{the local object if \code{remote_cytobands} is set to FALSE.}

\item{assembly, }{character, specify the genomic assembly. Can be either
"hg18", "hg19" or "hg38".}

\item{keep_str_end, }{logical, specify if intermediate columns (locus_start
and locus_end) must be kept or discarded.}
}
\description{
\code{genomic_locus} add the locus information to a \code{data.table}
containing CNV calls-like data
}
\details{
This function takes a \code{CNVresults} object and add a columns containing the
genomic locus information. By default the file containing the CytoBands location
for the specified assembly are downloaded from UCSC website, but also a local
object (as \code{data.table} or \code{data.frame}) can be used via the
argument \code{bands}.
}
