% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg18_chr_arm.R
\docType{data}
\name{hg18_chr_arms}
\alias{hg18_chr_arms}
\title{Chromosomal location of the genomic arms for the assembly hg18}
\format{
A \code{data.table} with four columns:
\describe{
  \item{chr}{chromosome name (1:24 format)}
  \item{arm_ID}{genomic arm name}
  \item{start}{genomic location of the arm start}
  \item{end}{genomic location of the arm end}
}
}
\usage{
hg18_chr_arms
}
\description{
Chromosomal location of the genomic arms for the assembly hg18
}
\keyword{datasets}
