% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_chr_st_en_centr.R
\docType{data}
\name{hg38_start_end_centromeres}
\alias{hg38_start_end_centromeres}
\title{Start, end and centrosome location of each chromosome for the assembly hg38}
\format{
A \code{data.table} with four columns:
\describe{
  \item{chr}{chromosome name (1:24 format)}
  \item{start}{start}
  \item{end}{end}
  \item{centromere}{centromere}
}
}
\usage{
hg38_start_end_centromeres
}
\description{
Start, end and centrosome location of each chromosome for the assembly hg38
}
\keyword{datasets}
