% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuno_regions.R
\name{immuno_regions}
\alias{immuno_regions}
\title{Retrieve genomic regions of consecutive immunoglobulin genes}
\usage{
immuno_regions(biotype = NULL, n_genes = 5, mart = NULL)
}
\arguments{
\item{biotype, }{vector of character, is used to filter the genes based on
\code{gene_biotype} from to \code{biomaRt::getBM}. By default all the
immunoglobulin biotypes al listed in Genecode
<https://www.gencodegenes.org/pages/biotypes.html>. If the user specify one
or more values only those are used.}

\item{n_genes, }{integer, number of minim consecutive genes required to
construct an "immunoglobulin region", default is 5.}

\item{mart, }{user provided mart (from \code{biomaRt::useMart} function). For
  older assemblies the user must manually retrieve the correct mart via
  \code{biomaRt}.

  \code{immuno_regions} is used to construct a \code{data.table} of genomic
  regions where all the consecutive gene have an immunoglobulin (or user
  specified). The output can be used as a blacklist in
  \code{\link{cleaning_filter}}.}
}
\value{
a list with two element, the first is a \code{data.table} containing
  the actual regions (to be passed to \code{\link{cleaning_filter}} for
  example), the second is the full genes list (for the required biotypes),
  again as \code{data.table}.
}
\description{
Retrieve genomic regions of consecutive immunoglobulin genes
}
