% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrr_trio_plot.R
\name{lrr_trio_plot}
\alias{lrr_trio_plot}
\title{Plot markers raw data in a CNV region for a trio}
\usage{
lrr_trio_plot(cnv, raw_path, sample_list, results)
}
\arguments{
\item{cnv, }{a \code{data.table} containig one single line of CNVs calling
results.}

\item{raw_path, }{\code{character}, the path where raw data processed by
\code{\link{read_NGS_raw}} or \code{\link{read_finalreport_raw}}.}

\item{sample_list, }{minimal cohort metadata, a \code{data.table} produced by the
function \code{\link{read_metadt}}.}

\item{results, }{a \code{data.table}, the output of
  \code{\link{read_results}}.

This function plot the raw data in the regions of interest in order to visually
confirm the presence of a good de novo CNV candidate.}
}
\value{
a \code{list} containing the \code{ggplot2} object.
}
\description{
Plot markers raw data in a CNV region for a trio
}
