% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_example.R
\docType{data}
\name{markers_examples}
\alias{markers_examples}
\title{Markers file for the runnable examples}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 115377 rows and 5 columns.
}
\usage{
markers_examples
}
\description{
Markers file for the runnable examples
}
\keyword{datasets}
