% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanti_22.R
\docType{data}
\name{quanti_22}
\alias{quanti_22}
\title{Chromosome 22 QuantiSNP example results for the runnable examples}
\format{
an object of \code{CNVresults} class.
}
\usage{
quanti_22
}
\description{
Chromosome 22 QuantiSNP example results for the runnable examples
}
\keyword{datasets}
