% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_final_report_raw.R
\name{read_finalreport_raw}
\alias{read_finalreport_raw}
\title{Read Illumina array raw data}
\usage{
read_finalreport_raw(
  DT_path,
  rds_path,
  pref,
  suff,
  sample_list,
  markers,
  chr_col = "Chr",
  pos_col = "Position",
  LRR_col = "Log R Ratio",
  BAF_col = "B Allele Freq"
)
}
\arguments{
\item{DT_path, }{character, the path to the directory with the raw data (i.e.
the splitted final reports)}

\item{rds_path, }{path to the directory where RDS should be stored.}

\item{pref, suff, }{eventual prefix an suffix (e.g. ".txt"). If not necessary
must be set to \code{NA}.}

\item{sample_list, }{a \code{data.table}, the output of
\code{\link{read_metadt}}.}

\item{markers, }{a \code{data.table}, the output of
\code{\link{read_NGS_intervals}} or \code{\link{read_finalreport_snps}},
depending on the initial data type.}

\item{chr_col, }{name of the column containing the chromosome information in
the input file. Default is \code{"Chr"}.}

\item{pos_col, }{name of the column containing the SNPs position information
in the input file. Default is \code{"Position"}.}

\item{LRR_col, }{name of the column containing the LRR information in the
input file. Default is \code{"Log R Ratio"}.}

\item{BAF_col, }{name of the column containing the BAF information in the
input file. Default is \code{"B Allele Freq"}.}
}
\value{
nothing, this function saves the results on disk.
}
\description{
\code{read_finalreport} handles inputs of data in FinalReport-like format
}
\details{
This function handles the input, pre-processing and temporary storage (as RDS
files) of the the markers-level raw data for each sample starting from
FinalReport-like files (any plain text file with columns header can be read).
Input must be one file per sample.
}
\examples{

read_finalreport_raw(DT_path = system.file("extdata", package = "CNVgears"),
                    pref = "Final_report_example", suff = ".txt",
                    rds_path = file.path("tmp_RDS"),
                    markers = markers_examples,
                    sample_list = cohort_examples[sample_ID == "NA12878", ])
}
