% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_final_report_snps.R
\name{read_finalreport_snps}
\alias{read_finalreport_snps}
\title{Read Illumina array raw data}
\usage{
read_finalreport_snps(
  DT_path,
  mark_ID_col = "SNP Name",
  chr_col = "Chr",
  pos_col = "Position"
)
}
\arguments{
\item{DT_path, }{path to the input file.}

\item{mark_ID_col, }{name of the column containing the SNP ID information in
the input file. Default is \code{"SNP Name"}.}

\item{chr_col, }{name of the column containing the chromosome information in
the input file. Default is \code{"Chr"}.}

\item{pos_col, }{name of the column containing the SNPs position information
in the input file. Default is \code{"Position"}.}
}
\value{
a \code{data.table}, will be of \code{Markers} class in future
  versions.
}
\description{
\code{read_finalreport} handles inputs of data in FinalReport-like format
}
\details{
This function is used to load data in FinalReport like format into a
\code{data.table} containing the SNPs markers information (i.e. chromosome
and position). The function expect a single file where each markers is
present one single time. Similar files are often required or produced by the
calling algorithm/pipeline, e.g. the PFB file in PennCNV can be used here.
}
\examples{
DT <- read_finalreport_snps(system.file("extdata", "SNP.pfb", package = "CNVgears"),mark_ID_col = "Name", chr_col = "Chr", pos_col = "Position")
}
