% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadt.R
\name{read_metadt}
\alias{read_metadt}
\title{Read sample file with minimal metadata}
\usage{
read_metadt(
  DT_path,
  sample_ID_col = "sample_ID",
  sex_col = "sex",
  role_col = "role",
  fam_ID_col = "fam_ID"
)
}
\arguments{
\item{DT_path}{path to the input file.}

\item{sample_ID_col}{name of the columns containing the sample ID in the original file;}

\item{sex_col}{name of the columns containing the sex information in the original file;}

\item{role_col}{name of the columns containing the role information ID in the original file;}

\item{fam_ID_col}{name of the columns containing the family ID in the original file;}
}
\value{
cohort metatadata object, a \code{data.table}. Will be of the
  \code{SampleList} class in future versions.
}
\description{
\code{read_medaDT} handles the input of the sample table (sampleID, sex, role,
famID) in a standardized format.
}
\details{
This function is needed in the first step of virtually every analysis. The input data
must have at least the following columns:
\itemize{
\item sample ID, self describing;
\item sex, ideally in 1/2 format, for males and females, however also "male"/"female"
  or "Male"/"Female" are accepted;
\item role, role of the sample in the family, either "father", "mother", "proband" or
"sibling";
\item family ID, self describing.
}
Actual name and order of the columns in the file is not relevant since they are
passed to the function via parameter.
Since the function in this package are optimized for family based studies, family ID and
role information for each sample are required, however if the user is interested only,
as an example, in CNVRs computation, genic content annotation or identification of calls
in IG regions and does not have such information "role" and famID can be "NA". Note that
doing so some functions won't be usable.
}
