% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NGS_raw.R
\name{read_NGS_raw}
\alias{read_NGS_raw}
\title{Read raw copyratio/LRR data for NGS intervals}
\usage{
read_NGS_raw(
  DT_path,
  rds_path,
  pref,
  suff,
  chr_col,
  start_col,
  end_col,
  raw_col,
  raw_type,
  sample_list,
  markers
)
}
\arguments{
\item{DT_path, }{path to the input file.}

\item{rds_path, }{path to the directory where RDS should be stored.}

\item{pref, suff, }{eventual prefix an suffix (e.g. ".txt"). If not necessary
must be set to \code{NA}.}

\item{chr_col, }{name of the column containing the chromosome information in
the input file.}

\item{start_col, }{name of the column containing the start information in the
input file.}

\item{end_col, }{name of the column containing the end information in the
input file.}

\item{raw_col, }{name of the column containing the marker-level raw data in
the input file.}

\item{raw_type, }{character, it describes the format in which the LRR-like
data is give. Can be either "log2R", "numeric_CN" or "copyratio", see
description.}

\item{sample_list, }{a \code{data.table}, the output of
\code{\link{read_metadt}}.}

\item{markers, }{a \code{data.table}, the output of
\code{\link{read_NGS_intervals}} or \code{\link{read_finalreport_snps}},
depending on the initial data type.}
}
\value{
nothing, this function saves the results on disk.
}
\description{
\code{read_NGS_raw}
}
\details{
This function handles the input, pre-processing and temporary storage (as RDS
files) of the the markers-level raw data for each sample. Those are mostly
required for CNVs inheritance detection. Any marker-level log2R-like measure
is supported, given that, for somatic chromosomes:
$$LRR = log2R = log2(CopyRatio) = log2(numeric_CN/2*)$$
Input must be one file per sample.
}
