% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telom_centrom.R
\name{telom_centrom}
\alias{telom_centrom}
\title{Generate blacklist for telomeric and centromeric regions}
\usage{
telom_centrom(DT_in, telom = TRUE, centrom = TRUE, region = 50000)
}
\arguments{
\item{DT_in, }{a \code{data.table} with start, and centromere position for
each chromosome. For the assemblies hg18, hg19 and hg38 this is provided in
the package.}

\item{telom, }{logical, should be telomeric blacklist be produced?}

\item{centrom, }{logical, should be centromeric blacklist be produced?}

\item{region, }{integer. How many basepairs large should the regions be? The
centromeric region will be twice this large.}
}
\value{
, a \code{data.table} that can be passed to
\code{\link{cleaning_filter}} as blacklist.
}
\description{
Generate blacklist for telomeric and centromeric regions
}
