% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmops_to_CNVresults.R
\name{cnmops_to_CNVresults}
\alias{cnmops_to_CNVresults}
\title{Convert cn.mops results into \code{CNVgears} format}
\usage{
cnmops_to_CNVresults(cnvs, sample_list, markers)
}
\arguments{
\item{cnvs}{cn.mops results}

\item{sample_list}{minimal cohort metadata, a \code{data.table} produced by the
function \code{\link{read_metadt}}.}

\item{markers}{a \code{data.table} containing the marker list, the output
\code{\link{read_finalreport_snps}} with \code{DT_type} set to "markers"
or \code{\link{read_NGS_intervals}}.}
}
\value{
the input object \code{cnvs} converted into \code{CNVresults}
}
\description{
Convert cn.mops results into \code{CNVgears} format
}
\examples{
\dontrun{
library(cn.mops)
data(cn.mops)
resCNMOPS <- cn.mops(XRanges)
resCNMOPS <- calcIntegerCopyNumbers(resCNMOPS)
resCNMOPS_cnvs <- cnvs(resCNMOPS)
cnmops_calls <- cnmops_to_CNVresults(resCNMOPS_cnvs)
}
}
