% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVRs_creation.R
\name{cnvrs_create}
\alias{cnvrs_create}
\title{Compute Copy Number Variable Regions (CNVRs)}
\usage{
cnvrs_create(cnvs, g_arms, prop = 0.3)
}
\arguments{
\item{g_arms, }{a data.table containing the genomic location of the genomic
arms. For the assemblies hg38 and hg19 it is provided by the package.}

\item{prop, }{reciprocal overlap proportion, default 0.3 (30%).}

\item{cnv, }{a data.table like the one produced by\code{\link{read_results}},
typically the results of \code{\link{inter_res_merge}}.}

\item{inner_outer, }{specify whether inner or outer start/end should be used. If
NA the function assume the "start" "end" columns are present.

lorem ipsum etc etc}
}
\value{

}
\description{
Compute Copy Number Variable Regions (CNVRs)
}
