% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_results_compare_merge.R
\name{inter_res_merge}
\alias{inter_res_merge}
\title{Combine the results from multiple methods in a single object}
\usage{
inter_res_merge(res_list, sample_list, g_arms, prop = 0.3, inner_outer = NA)
}
\arguments{
\item{g_arms}{}
}
\value{

}
\description{
Combine the results from multiple methods in a single object
}
