% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_example.R
\docType{data}
\name{markers_examples}
\alias{markers_examples}
\title{Markers file for the runnable examples
It comes from CNV calling on array data from a subset of the 1000 Genomes
project, see vignettes for more details.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 115377 rows and 5 columns.
}
\usage{
markers_examples
}
\description{
Markers file for the runnable examples
It comes from CNV calling on array data from a subset of the 1000 Genomes
project, see vignettes for more details.
}
\keyword{datasets}
