% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_results_merging.R
\name{merge_calls}
\alias{merge_calls}
\title{Merge adjacent CNV with equal Copy Number}
\usage{
merge_calls(DT_in, prop = 0.3)
}
\arguments{
\item{DT_in}{a \code{data.table} in the format of
\code{\link{read_results}} output.}

\item{prop}{numeric value that multiplied by the length of a call gives
the range where adjacent calls are searched.}
}
\value{
a \code{CNVresults}, \code{DT_in} processing result.
}
\description{
\code{merge_calls} screen the segments of a sample and merge adjacent calls
with equal GT if close enough
}
\details{
This function takes a \code{CVNResults} object and try to merge adjacent calls
(having equal GT and being close
enough). It is designed to process one sample at the time and is integrated
in the function \code{\link{read_results}}. It is not suggested to use it
stand-alone. The merging is done starting from the larger calls, on adjacent
call with equal GT in a range of +/- \code{length*prop}, trough several
iterations until no more calls can be merged.
}
\examples{
# with merge
DT <- read_results(DT_path = system.file("extdata", "chrs_14_22_cnvs_penn.txt", package = "CNVgears"), res_type = "file", DT_type = "TSV/CSV", chr_col = "chr", start_col = "posStart", end_col = "posEnd", CN_col = "CN", samp_ID_col = "Sample_ID", sample_list = cohort_examples, markers = markers_examples, method_ID = "P")
# without
DT <- read_results(do_merge = FALSE, DT_path = system.file("extdata", "chrs_14_22_cnvs_penn.txt", package = "CNVgears"), res_type = "file", DT_type = "TSV/CSV", chr_col = "chr", start_col = "posStart", end_col = "posEnd", CN_col = "CN", samp_ID_col = "Sample_ID", sample_list = cohort_examples, markers = markers_examples, method_ID = "P")

}
