% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penn_22.R
\docType{data}
\name{penn_22}
\alias{penn_22}
\title{Chromosome 22 PennCNV example results for the runnable examples}
\format{
an object of \code{CNVresults} class.
}
\usage{
penn_22
}
\description{
Chromosome 22 PennCNV example results for the runnable examples
}
\keyword{datasets}
