% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanti_22.R
\docType{data}
\name{quanti_22}
\alias{quanti_22}
\title{Chromosome 22 QuantiSNP example results for the runnable examples.
The results of CNV calling using the program QuantiSNP on array data from
a subset of the 1000 Genomes project, see vignettes for more details. Only
chromosome 22 was kept in order to keep the object small.}
\format{
an object of \code{CNVresults} class.
}
\usage{
quanti_22
}
\description{
Chromosome 22 QuantiSNP example results for the runnable examples.
The results of CNV calling using the program QuantiSNP on array data from
a subset of the 1000 Genomes project, see vignettes for more details. Only
chromosome 22 was kept in order to keep the object small.
}
\keyword{datasets}
