% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{Read CNVs calling or semgentation results}
\usage{
read_results(
  DT_path,
  res_type,
  DT_type,
  pref = NA,
  suff = NA,
  sample_list,
  markers,
  chr_col,
  start_col,
  end_col,
  CN_col,
  samp_ID_col,
  end_vcf = "END",
  CN_vcf = "CN",
  do_merge = TRUE,
  merge_prop = 0.5,
  method_ID
)
}
\arguments{
\item{DT_path, }{path to the directory containing the individual files, if
\code{res_type} is set to "directory" or to the single file if \code{res_type}
is set to "file".}

\item{res_type, }{can be either "directory" or "file", indicates whether the
function must expect a single file for all samples or one file per sample.}

\item{DT_type, }{can be either "VCF" or "TSV/CSV", indicate the file type.}

\item{pref, suff, }{eventual prefix an suffix (e.g. ".txt") to the files to be used
when \code{res_type} is set to "directory". If not necessary must be set to
\code{NA}.}

\item{sample_list, }{minimal cohort metadata, a \code{data.table} produced by the
function \code{\link{read_metadt}}.}

\item{markers, }{a \code{data.table} containing the marker list, the output
\code{\link{read_finalreport_snps}} with \code{DT_type} set to "markers"
or \code{\link{read_NGS_intervals}}.}

\item{chr_col, }{name of the column containing the chromosome information
in the input data.}

\item{start_col, }{name of the column containing the start information in
the input data.}

\item{end_col, }{name of the column containing the end information in the
input data.}

\item{CN_col, }{name of the column containing the Copy Number information
in the input data.}

\item{samp_ID_col, }{name of the column containing the sample ID information in
the input file, required if \code{res_type} is set to "file".}

\item{end_vcf, }{name of the field containing the segment end information in the
VCF file(s), passed to the function \code{\link{read_vcf}}.}

\item{CN_vcf, }{name of the field containing the segment copy number
information in the VCF file(s), passed to the function \code{\link{read_vcf}}.}

\item{do_merge, }{logical, indicates whether the function
\code{\link{merge_calls}} should be automatically called for each sample
(strongly suggested).}

\item{merge_prop}{minimum reciprocal overlap proportion in order to merge.}

\item{method_ID, }{character identifying the method (algorithms/pipeline), one
letter code is strongly encouraged (e.g. "P" for PennCNV and "M" for GATK
ModSeg). Numeric are converted to character.}
}
\value{
a \code{CNVresults} object.
}
\description{
\code{read_results} takes the results of a CNVs calling pipeline and return
them in a standardized object.
}
\details{
This function aims to convert a variety of possible types of CNVs
calling/segmentation pipelines and/or algorithms results into a standardized
format in order to easily integrate with the other functions in this
package. Currently two main files type and two main file-organization structures
are considered, for a total of four generic situations:
\itemize{
\item VCF files, one per sample (e.g. the results of GATK gCNV pipeline);
\item VCF file, all sample of a cohort in the same file (not yet fully implemented);
\item TSV/CSV file, one file per sample (e.g. the results of GATK ModSeg pipeline,
  or the results of running "manually" PennCNV);
\item TSV/CSV file, all samples of a cohort in the same file (e.g. the results
  of EnsembleCNV).
}
If multiple files containing results for multiple samples are present (e.g. the
results of PennCNV joint calling on trios) at the moment it is recommended that
the user concatenated those individual file in a single one prior loading them
with \code{read_results}.
Note that any line occurring before the columns header are automatically skipped
by \code{fread}.
}
