% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_VCF.R
\name{read_vcf}
\alias{read_vcf}
\title{Convert a VCF file of genomics segments into a \code{data.table}}
\usage{
read_vcf(DT_path, end_vcf = "END", CN_vcf = "CN", samples = NA, explore = F)
}
\arguments{
\item{DT_path, }{path to the file.}

\item{end_vcf, }{name of the field containing the segment end information.}

\item{CN_vcf, }{name of the field containing the segment copy number
information.}

\item{samples, }{NA by default, if a character vector is providend is used to
identify and select samples in a VCF containg multiple ones.}

\item{explore, }{logic, \code{FALSE} by default. If \code{TRUE} the file in
\code{DT_path} is not loaded, insted, several infos about the VCF fields
are printed.}
}
\description{
\code{read_vcf} read a VCF file into a \code{data.table}
}
\details{
This function use \code{readVcf} from \code{VariantAnnotation} to read VCF
files, then it select only the necessary columns (for the purpose of CNVs
calling results analysis) and convert it to a \code{data.table}. Can also be
usewd to check the names of the necessary fields (end and copy number) if not
already known, using the parameter \code{explore}. By default it expect a
file containing data for a single sample (e.g. the results of gCNV from
GATK), but it can process files containing multiple samples if a charater
vector containig the IDs is given to the parameter \code{samples}.
}
